#8 매퍼 인터페이스와 레포지토리 메소드에서 잘못된 타입을 반환하고 있어서 수정했습니다(int --> float)

+++ src/main/java/com/tcp/toeflserver/place/PlaceRepository.java
@@ -15,7 +15,7 @@ public class PlaceRepository {
         placeMapper.updatePlace(place);
     }
 
-    int selectPlaceReviewScoreAverage(String placeId){
+    float selectPlaceReviewScoreAverage(String placeId){
         return placeMapper.selectPlaceReviewAverageScore(placeId);
     }
 }
